requirements.txt
# Core Dependencies
# ==================

# Encryption and Security
cryptography>=41.0.0          # AES-256 encryption, PBKDF2 key derivation, secure random generation
bcrypt>=4.0.0                  # Secure password hashing for user authentication

# GUI Framework
customtkinter>=5.2.0           # Modern, customizable GUI framework with Windows 11 styling
pillow>=10.0.0                 # Image processing for GUI icons and themes

# Database
# sqlite3 is built into Python, no external dependency needed

# Utilities
pyperclip>=1.8.0              # Clipboard operations for copy/paste functionality
zxcvbn>=4.4.28                # Advanced password strength checking and analysis
python-dateutil>=2.8.0        # Date parsing and manipulation for backup timestamps
python-dotenv>=1.0.0          # Environment variable loading from .env files

# Web Interface
flask>=2.3.0                  # Lightweight web framework for web interface
flask-session>=0.5.0          # Session management for web interface security
flask-wtf>=1.1.0              # CSRF protection and form validation
flask-limiter>=3.5.0          # Rate limiting for brute force protection
wtforms>=3.0.0                # Form validation and rendering
jinja2>=3.1.0                 # Template engine for HTML generation
werkzeug>=2.3.0               # WSGI utilities and development server
itsdangerous>=2.1.0           # Cryptographic signing for session cookies

# Cloud Sync (Optional Features)
requests>=2.31.0              # HTTP library for cloud storage API interactions
google-api-python-client>=2.100.0  # Google Drive API integration (optional)
dropbox>=11.36.0              # Dropbox API integration (optional)

# Testing and Development
pytest>=7.4.0                # Testing framework for unit and integration tests
pytest-cov>=4.1.0            # Test coverage reporting
black>=23.7.0                 # Code formatting (development)
flake8>=6.0.0                 # Linting and style checking (development)

# Import/Export Support
pandas>=2.1.0                 # Data manipulation for CSV import/export
openpyxl>=3.1.0              # Excel file support for import/export
lxml>=4.9.0                   # XML processing for various password manager formats

# System Integration
psutil>=5.9.0                 # System information for dependency checking
packaging>=23.1               # Version parsing for dependency verification
colorama>=0.4.0               # Cross-platform colored terminal text (Windows compatibility)
chardet>=5.0.0                # Character encoding detection for CSV import

# Testing and Development
pytest>=7.4.0                 # Testing framework (alternative to unittest)
coverage>=7.3.0               # Code coverage measurement
pytest-cov>=4.1.0            # Coverage plugin for pytest
pytest-mock>=3.11.0          # Mocking plugin for pytest
selenium>=4.15.0              # Web browser automation for end-to-end tests (optional)
mock>=5.1.0                   # Mock object library for testing

# Executable Building
pyinstaller>=6.0.0           # Creates standalone executables for distribution
